Split dashboard API provisioning handlers
